requirements.txt
tensorflow==2.13.0
keras==2.13.1
pandas==2.0.3
numpy==1.24.3
jupyter==1.0.0
notebook==6.5.4
tqdm==4.65.0
opencv-contrib-python==4.8.1.78
scikit-learn==1.3.0
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-cors==4.0.0
reportlab==4.0.4
cryptography==41.0.4
requests==2.31.0
huggingface_hub==0.16.4
azure-ai-inference==0.1.0
google-generativeai==0.3.0
Pillow==10.0.0
python-dateutil==2.8.2
werkzeug==2.3.7
